+++ exporter/prometheusexporter/collector.go
@@ -88,9 +88,10 @@ func (c *collector) getMetricMetadata(metric pdata.Metric, labels pdata.StringMa
 	keys := make([]string, 0, labels.Len())
 	values := make([]string, 0, labels.Len())
 
-	labels.ForEach(func(k string, v string) {
+	labels.Range(func(k string, v string) bool {
 		keys = append(keys, sanitize(k))
 		values = append(values, v)
+		return true
 	})
 
 	return prometheus.NewDesc(